moved hex string to bytes function into utilities
